Move login and register features  to Auth Feature
