+++ lib/core/di/dependency_injection.dart
@@ -1,8 +1,8 @@
 import 'package:dio/dio.dart';
 import 'package:get_it/get_it.dart';
 
-import '../../features/login/data/repos/login_repo.dart';
-import '../../features/login/logic/cubit/login_cubit.dart';
+import '../../features/auth/login/data/repos/login_repo.dart';
+import '../../features/auth/login/logic/cubit/login_cubit.dart';
 import '../networking/api_service/api_service.dart';
 import '../networking/dio/dio_factory.dart';
