+++ lib/core/networking/api_service/api_service.dart
@@ -1,8 +1,8 @@
 import 'package:dio/dio.dart';
 import 'package:retrofit/retrofit.dart';
 
-import '../../../features/login/data/models/login_request_body.dart';
-import '../../../features/login/data/models/login_response.dart';
+import '../../../features/auth/login/data/models/login_request_body.dart';
+import '../../../features/auth/login/data/models/login_response.dart';
 import '../api_constants.dart';
 
 part 'api_service.g.dart';
