+++ lib/features/auth/login/ui/login_screen.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 
-import '../../../core/di/dependency_injection.dart';
-import '../../../core/helpers/extensions/extensions.dart';
+import '../../../../core/di/dependency_injection.dart';
+import '../../../../core/helpers/extensions/extensions.dart';
 import '../logic/cubit/login_cubit.dart';
 import 'widgets/already_have_an_account.dart';
 import 'widgets/login_form.dart';
